Refactored copy task methods from database
