+++ sayn/tasks/copy.py
@@ -1,5 +1,7 @@
 import logging
 
+from sqlalchemy.sql import select, func, or_, text
+
 from .task import TaskStatus
 from .sql import SqlTask
 
@@ -49,13 +51,13 @@ class CopyTask(SqlTask):
                 res = self.db.select(self.compiled[step])
                 res = list(res[0].values())[0]
                 if res is not None:
-                    last_incremental_value = {'incremental_value': res}
+                    last_incremental_value = {"incremental_value": res}
 
             step = "get_data"
             data = self.source_db.select(self.compiled[step], last_incremental_value)
 
             if len(data) == 0:
-                logging.debug('Nothing to load')
+                logging.debug("Nothing to load")
                 return self.success()
 
             step = "create_load_table"
@@ -112,7 +114,7 @@ class CopyTask(SqlTask):
         # Type of materialisation
         self.delete_key = self._pop_property("delete_key")
         self.incremental_key = self._pop_property("incremental_key")
-        
+
         if (self.delete_key is None) != (self.incremental_key is None):
             return self.failed(
                 f'Incremental copy requires both "delete_key" and "incremental_key"'
@@ -122,17 +124,19 @@ class CopyTask(SqlTask):
 
     def _setup_source(self):
         # Source property indicating the table this will create
-        source = self._pop_property(
-            "source", default={"schema": None}
-        )
+        source = self._pop_property("source", default={"schema": None})
 
-        self.schema = source.pop('schema', None)
+        self.schema = source.pop("schema", None)
         if self.schema is not None and isinstance(self.schema, str):
             self.schema = self.compile_property(self.schema)
         else:
             return self.failed('Optional property "schema" must be a string')
 
-        if set(source.keys()) != set(["db", "table"]) or source['db'] is None or source['table'] is None:
+        if (
+            set(source.keys()) != set(["db", "table"])
+            or source["db"] is None
+            or source["table"] is None
+        ):
             return self.failed(
                 'Source requires "table" and "db" fields. Optional field: "schema".'
             )
@@ -141,7 +145,9 @@ class CopyTask(SqlTask):
             self.source_table = self.compile_property(source.pop("table"))
 
         if source_db_name not in self.sayn_config.dbs:
-            return self.failed(f'{source_db_name} is not a valid vallue for "db" in "source"')
+            return self.failed(
+                f'{source_db_name} is not a valid vallue for "db" in "source"'
+            )
         else:
             self.source_db = self.sayn_config.dbs[source_db_name]
 
@@ -152,7 +158,7 @@ class CopyTask(SqlTask):
             self.source_table,
             self.source_schema,
             columns=[c["name"] for c in self.ddl["columns"]],
-            required_existing=True
+            required_existing=True,
         )
 
         if self.source_table_def is None or not self.source_table_def.exists():
@@ -180,7 +186,7 @@ class CopyTask(SqlTask):
         )
 
         if self.destination_table_def is None:
-            return self.failed('Error detecting destination table')
+            return self.failed("Error detecting destination table")
 
         if not self.destination_table_def.exists():
             # Full load:
@@ -192,7 +198,7 @@ class CopyTask(SqlTask):
             )
             self.compiled = {
                 "last_incremental_value": None,
-                "get_data": self.source_db.get_data(
+                "get_data": self.get_data(
                     self.source_table,
                     self.source_schema,
                     [c["name"] for c in self.ddl["columns"]],
@@ -211,7 +217,7 @@ class CopyTask(SqlTask):
             )
             self.compiled = {
                 "last_incremental_value": None,
-                "get_data": self.source_db.get_data(
+                "get_data": self.get_data(
                     self.source_table,
                     self.source_schema,
                     [c["name"] for c in self.ddl["columns"]],
@@ -234,10 +240,10 @@ class CopyTask(SqlTask):
                 self.load_table, self.load_schema, self.ddl, replace=True
             )
             self.compiled = {
-                "last_incremental_value": self.db.get_max_value(
+                "last_incremental_value": self.get_max_value(
                     self.table, self.schema, self.incremental_key
                 ),
-                "get_data": self.source_db.get_data(
+                "get_data": self.get_data(
                     self.source_table,
                     self.source_schema,
                     [c["name"] for c in self.ddl["columns"]],
@@ -256,3 +262,86 @@ class CopyTask(SqlTask):
             raise ValueError("Sayn error in copy")
 
         return TaskStatus.READY
+
+    # Execution steps
+
+    def get_max_value(self, table, schema, column):
+        table_def = self.db.get_table(table, schema)
+        return select([func.max(table_def.c[column])]).where(
+            table_def.c[column] != None
+        )
+
+    def get_data(self, table, schema, columns, incremental_key=None):
+        src = self.source_db.get_table(table, schema, columns)
+        q = select([src.c[c] for c in columns])
+        if incremental_key is not None:
+            q = q.where(
+                or_(
+                    src.c[incremental_key] == None,
+                    src.c[incremental_key] > text(":incremental_value"),
+                )
+            )
+        return q
+
+    # def copy_from_table(
+    #     self, src, dst_table, dst_schema=None, columns=None, incremental_column=None
+    # ):
+    #     # 1. Check source table
+    #     if not src.exists():
+    #         raise DatabaseError(f"Source table {src.fullname} does not exists")
+
+    #     if columns is None:
+    #         columns = [c.name for c in src.columns]
+    #         if incremental_column is not None and incremental_column not in src.columns:
+    #             raise DatabaseError(
+    #                 f"Incremental column {incremental_column} not in source table {src.fullname}"
+    #             )
+    #     else:
+    #         if incremental_column is not None and incremental_column not in columns:
+    #             columns.append(incremental_column)
+
+    #         for column in columns:
+    #             if column not in src.columns:
+    #                 raise DatabaseError(
+    #                     f"Specified column {column} not in source table {src.fullname}"
+    #                 )
+
+    #     # 2. Get incremental value
+    #     table = self.get_table(dst_table, dst_schema, columns=columns)
+
+    #     if incremental_column is not None and not dst_table.exists():
+    #         if incremental_column not in table.columns:
+    #             raise DatabaseError(
+    #                 f"Incremental column {incremental_column} not in destination table {table.fullname}"
+    #             )
+    #         incremental_value = (
+    #             select([func.max(table.c[incremental_column])])
+    #             .where(table.c[incremental_column] != None)
+    #             .execute()
+    #             .fetchone()[0]
+    #         )
+    #     elif table is None:
+    #         # Create table
+    #         table = Table(dst_table, self.metadata, schema=dst_schema)
+    #         for column in columns:
+    #             table.append_column(src.columns[column].copy())
+    #         table.create()
+    #         incremental_value = None
+    #     else:
+    #         incremental_value = None
+
+    #     # 3. Get data
+    #     if incremental_value is None:
+    #         where_cond = True
+    #     else:
+    #         where_cond = or_(
+    #             src.c[incremental_column] == None,
+    #             src.c[incremental_column] > incremental_value,
+    #         )
+    #     query = select([src.c[c] for c in columns]).where(where_cond)
+    #     data = query.execute().fetchall()
+
+    #     # 4. Load data
+    #     return self.load_data(
+    #         [dict(zip([str(c.name) for c in query.c], row)) for row in data], table
+    #     )
